requirements.txt
requests
python-telegram-bot
pyrogram
tgcrypto
